requirements.txt
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
Pillow==6.2.2
torch==1.7.0+cpu
torchvision==0.8.1+cpu
gunicorn==20.0.4
itsdangerous==1.1.0
certifi==2020.12.5